requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
aiofiles=24.1.0=pypi_0
aiohappyeyeballs=2.5.0=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.8.0=pypi_0
asgiref=3.8.1=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.1.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.3.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
blinker=1.9.0=pypi_0
build=1.2.2.post1=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
cachetools=5.5.2=pypi_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.4.1=pypi_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.6.3=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=pypi_0
cryptography=44.0.2=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.8.11=py312h5da7b33_0
decorator=5.2.1=pyhd8ed1ab_0
deprecated=1.2.18=pypi_0
dirtyjson=1.0.8=pypi_0
distro=1.9.0=pypi_0
durationpy=0.9=pypi_0
effdet=0.4.1=pypi_0
emoji=2.14.1=pypi_0
et-xmlfile=2.0.0=pypi_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.6.4=h8ddb27b_0
fastapi=0.115.11=pypi_0
filelock=3.17.0=pypi_0
filetype=1.2.0=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2025.3.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
google-api-core=2.24.2=pypi_0
google-auth=2.38.0=pypi_0
google-cloud-vision=3.10.0=pypi_0
googleapis-common-protos=1.69.1=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.71.0=pypi_0
grpcio-status=1.71.0=pypi_0
h11=0.14.0=pypi_0
html5lib=1.1=pypi_0
httpcore=1.0.7=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface-hub=0.29.3=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.5.2=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.29.5=pyh4bbf305_0
ipython=9.0.2=pyhca29cf9_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
jiter=0.9.0=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=py312haa95532_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=hdf4eb48_0
kubernetes=32.0.1=pypi_0
langchain=0.3.20=pypi_0
langchain-community=0.3.19=pypi_0
langchain-core=0.3.43=pypi_0
langchain-openai=0.3.8=pypi_0
langchain-text-splitters=0.3.6=pypi_0
langdetect=1.0.9=pypi_0
langsmith=0.3.13=pypi_0
layoutparser=0.3.4=pypi_0
libffi=3.4.4=hd77b12b_1
libsodium=1.0.20=hc70643c_0
llama-cloud=0.1.14=pypi_0
llama-cloud-services=0.6.5=pypi_0
llama-index=0.12.23=pypi_0
llama-index-agent-openai=0.4.6=pypi_0
llama-index-cli=0.4.1=pypi_0
llama-index-core=0.12.23.post2=pypi_0
llama-index-embeddings-openai=0.3.1=pypi_0
llama-index-indices-managed-llama-cloud=0.6.8=pypi_0
llama-index-llms-ollama=0.5.3=pypi_0
llama-index-llms-openai=0.3.25=pypi_0
llama-index-multi-modal-llms-openai=0.4.3=pypi_0
llama-index-program-openai=0.3.1=pypi_0
llama-index-question-gen-openai=0.3.0=pypi_0
llama-index-readers-file=0.4.6=pypi_0
llama-index-readers-llama-parse=0.4.0=pypi_0
llama-parse=0.6.4.post1=pypi_0
lxml=5.3.1=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.26.1=pypi_0
matplotlib=3.10.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mmh3=5.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=1.30.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
olefile=0.47=pypi_0
ollama=0.4.7=pypi_0
omegaconf=2.3.0=pypi_0
onnx=1.17.0=pypi_0
onnxruntime=1.21.0=pypi_0
openai=1.65.5=pypi_0
opencv-python=4.11.0.86=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.4.1=ha4e3fda_0
opentelemetry-api=1.30.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.30.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.30.0=pypi_0
opentelemetry-instrumentation=0.51b0=pypi_0
opentelemetry-instrumentation-asgi=0.51b0=pypi_0
opentelemetry-instrumentation-fastapi=0.51b0=pypi_0
opentelemetry-proto=1.30.0=pypi_0
opentelemetry-sdk=1.30.0=pypi_0
opentelemetry-semantic-conventions=0.51b0=pypi_0
opentelemetry-util-http=0.51b0=pypi_0
orjson=3.10.15=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pdf2image=1.17.0=pypi_0
pdfminer-six=20231228=pypi_0
pdfplumber=0.11.5=pypi_0
pi-heif=0.21.0=pypi_0
pickleshare=0.7.5=pyhd8ed1ab_1004
pikepdf=9.5.2=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py312haa95532_0
platformdirs=4.3.6=pyhd8ed1ab_1
portalocker=3.1.1=pypi_0
posthog=3.19.1=pypi_0
prompt-toolkit=3.0.50=pyha770c72_0
propcache=0.3.0=pypi_0
proto-plus=1.26.1=pypi_0
protobuf=5.29.3=pypi_0
psutil=5.9.0=py312h827c3e9_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=19.0.1=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycocotools=2.0.8=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydantic-settings=2.8.1=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pypandoc=1.15=pypi_0
pyparsing=3.2.1=pypi_0
pypdf=5.3.1=pypi_0
pypdfium2=4.30.1=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
pyreadline3=3.5.4=pypi_0
python=3.12.9=h14ffc60_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-docx=1.1.2=pypi_0
python-dotenv=1.0.1=pypi_0
python-iso639=2025.2.18=pypi_0
python-magic=0.4.27=pypi_0
python-multipart=0.0.20=pypi_0
python-oxmsg=0.0.2=pypi_0
python-pptx=1.0.2=pypi_0
pytz=2025.1=pypi_0
pywin32=308=py312h5da7b33_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py312h5da7b33_0
rapidfuzz=3.12.2=pypi_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.23.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.5.3=pypi_0
scipy=1.15.2=pypi_0
setuptools=75.8.0=py312haa95532_0
shellingham=1.5.4=pypi_0
six=1.17.0=pyhd8ed1ab_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.38=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.46.1=pypi_0
streamlit=1.43.1=pypi_0
striprtf=0.0.26=pypi_0
sympy=1.13.1=pypi_0
tenacity=9.0.0=pypi_0
tiktoken=0.9.0=pypi_0
timm=1.0.15=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.21.0=pypi_0
toml=0.10.2=pypi_0
torch=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tornado=6.4.2=py312h827c3e9_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.49.0=pypi_0
typer=0.15.2=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.0=pypi_0
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025.1=pypi_0
ucrt=10.0.22621.0=h57928b3_1
unstructured=0.16.10=pypi_0
unstructured-client=0.31.1=pypi_0
unstructured-inference=0.8.1=pypi_0
unstructured-pytesseract=0.3.15=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
vc=14.42=haa95532_4
vc14_runtime=14.42.34438=hfd919c2_24
vs2015_runtime=14.42.34438=h7142326_24
watchdog=6.0.0=pypi_0
watchfiles=1.0.4=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py312haa95532_0
wrapt=1.17.2=pypi_0
xlrd=2.0.1=pypi_0
xlsxwriter=3.2.2=pypi_0
xz=5.6.4=h4754444_1
yarl=1.18.3=pypi_0
zeromq=4.3.5=ha9f60a1_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h8cc25b3_1
zstandard=0.23.0=pypi_0
